+++ zakozhurnikova.kristina/S5/inputMap.cpp
@@ -1,14 +1,14 @@
 #include "inputMap.hpp"
 #include <istream>
 
-void zakozhurnikova::inputMap(std::istream& in, BinarySearchTree< int, std::string >& map)
+void zakozhurnikova::inputMap(std::istream& in, BinarySearchTree< long long, std::string >& map)
 {
   ScopeGuard guard(in);
   int key = 0;
   std::string value;
   while (in)
   {
-    BinarySearchTree< int, std::string > tree;
+    BinarySearchTree< long long, std::string > tree;
     while (in >> key >> value)
     {
       tree.push(key, value);
